requirements.txt
# Core Python packages for CPAS4
PyQt6>=6.4.0
langchain>=0.0.340
langgraph>=0.0.40
langchain-community>=0.0.10
pymongo>=4.5.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.1
requests>=2.31.0
ollama>=0.1.6
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
openai>=1.3.0
tiktoken>=0.5.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
torch>=2.1.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
python-dateutil>=2.8.2
pytz>=2023.3
schedule>=1.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
uvicorn>=0.24.0
fastapi>=0.104.1
jinja2>=3.1.2
aiofiles>=23.2.1
websockets>=12.0
